Split session recording manager into focused modules
- move capture orchestration helpers into `recording_capture.py`
- move device selection/state helpers into `recording_device_selection.py`
- move recording state lifecycle helpers into `recording_lifecycle.py`
- move unlock/refresh authorization helpers into `recording_unlock.py`
